# server/docs/swagger.yml
openapi: 3.0.3
info:
  title: Barebone
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /auths:
    get:
      tags:
        - Auth
      summary: list Auths
      description: Retrieve a list of Auths
      operationId: list-auths-get
      parameters:
        - name: search
          in: query
          description: Search query string
          required: false
          schema:
            type: string
            example: ""
        - name: order_by
          in: query
          description: Field to order by
          required: false
          schema:
            type: string
            example: created_at
        - name: order
          in: query
          description: Order direction
          required: false
          schema:
            type: string
            enum:
              - DESC
              - ASC
            example: DESC
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Number of results per page
          required: false
          schema:
            type: integer
            default: 10
        - name: trash
          in: query
          description: Include trashed items
          required: false
          schema:
            type: boolean
            default: false
        - name: where_clause
          in: query
          description: Additional filtering criteria
          required: false
          schema:
            type: object
            example: {}
        - name: select
          in: query
          description: Specific fields to select
          required: false
          schema:
            type: string
            example: ""
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Entities retrieved successfully
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Auth"
                  info:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 10
                      trash:
                        type: boolean
                        example: false
                      order_by:
                        type: string
                        example: created_at
                      order:
                        type: string
                        example: DESC
                      search:
                        type: string
                        example: ""
                      where_clause:
                        type: object
                        example: {}
                      select:
                        type: string
                        example: ""
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      security:
        - auth:
            - list.auths
...
  /auth:
    get:
      tags:
        - Auth
      summary: list Auths
      description: Retrieve a list of Auths
      operationId: list-auth-get
      parameters:
        - name: search
          in: query
          description: Search query string
          required: false
          schema:
            type: string
            example: ""
        - name: order_by
          in: query
          description: Field to order by
          required: false
          schema:
            type: string
            example: created_at
        - name: order
          in: query
          description: Order direction
          required: false
          schema:
            type: string
            enum:
              - DESC
              - ASC
            example: DESC
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Number of results per page
          required: false
          schema:
            type: integer
            default: 10
        - name: trash
          in: query
          description: Include trashed items
          required: false
          schema:
            type: boolean
            default: false
        - name: where_clause
          in: query
          description: Additional filtering criteria
          required: false
          schema:
            type: object
            example: {}
        - name: select
          in: query
          description: Specific fields to select
          required: false
          schema:
            type: string
            example: ""
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Entities retrieved successfully
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Auth"
                  info:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 10
                      trash:
                        type: boolean
                        example: false
                      order_by:
                        type: string
                        example: created_at
                      order:
                        type: string
                        example: DESC
                      search:
                        type: string
                        example: ""
                      where_clause:
                        type: object
                        example: {}
                      select:
                        type: string
                        example: ""
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      security:
        - auth:
            - list.auth
...
components:
  schemas:
...
    Auth:
      properties:
        title:
          minLength: 3
          type: string
      type: object
      required:
        - title
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header